requirements.txt
biopython==1.86
pandas==2.3.3
matplotlib==3.9.0
seaborn==0.13.2
numpy==1.26.4
